refactor(presentation): 키워드 칩을 TypographyLabel 서브클래스로 단순화
- KeywordChipView(UIView 래퍼) → KeywordChipLabel(TypographyLabel 서브클래스)
- drawText(in:) 및 intrinsicContentSize 오버라이드로 패딩 처리해 래퍼 뷰 제거
- KeywordsCell 및 Constant 주석의 참조명 동기화
